Moved classes in SplooshModel.cs into separate files
